refactor: streamline shape selection logic and move distance functions
